requirements.txt
absl-py==1.2.0
alembic==1.7.3
altair==4.1.0
amqp==5.1.1
anchor==0.4.0
anyio==3.3.2
appdirs==1.4.4
argon2-cffi==21.1.0
args==0.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
audioread==2.1.9
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==4.0.1
bleach==4.1.0
blinker==1.4
blis==0.7.6
bokeh==2.4.0
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.5.30
certipy==0.1.3
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
clang==5.0
click==7.1.2
clint==0.5.1
cloudpickle==2.0.0
colorama==0.4.4
colorcet==2.0.6
colour==0.1.5
combo==0.1.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
cupshelpers==1.0
cvxopt==1.2.7
cycler==0.10.0
cymem==2.0.6
Cython==0.29.27
dask==2021.9.1
dataclasses==0.6
datashader==0.13.0
datashape==0.5.2
dbus-python==1.2.16
debtcollector==2.5.0
debugpy==1.5.0
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.4
distlib==0.3.3
distributed==2021.9.1
distro==1.6.0
distro-info==0.0.0
dm-tree==0.1.6
dnspython==2.2.1
docopt==0.6.2
dtreeviz==1.3.1
entrypoints==0.3
et-xmlfile==1.1.0
eventlet==0.33.1
extratools==0.8.2.1
filelock==3.3.0
fire==0.4.0
fixtures==4.0.1
flatbuffers==1.12
fsspec==2021.10.0
funcy==1.16
future==0.18.2
futurist==2.4.1
gast==0.4.0
gensim==4.1.2
gin-config==0.5.0
gitdb==4.0.7
GitPython==3.1.24
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.33.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
graphviz==0.17
greenlet==1.1.2
grpcio==1.47.0
h5py==3.7.0
HeapDict==1.0.1
holoviews==1.14.6
htmlmin==0.1.12
httplib2==0.19.1
hwinfo==0.1.0
hyperlink==21.0.0
idna==3.2
ImageHash==4.2.1
imageio==2.9.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
incremental==21.3.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==1.1.0
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonpickle==2.0.0
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyter-telemetry==0.1.0
jupyterhub==1.4.2
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
jupyterthemes==0.20.0
kaggle==1.5.12
keras==2.9.0
keras-nightly==2.9.0.dev2022021808
Keras-Preprocessing==1.1.2
keyring==23.2.1
kiwisolver==1.3.2
kmodes==0.11.1
kombu==5.2.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.14
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
lesscpy==0.15.0
libclang==14.0.1
librosa==0.8.1
lime==0.2.0.1
llvmlite==0.36.0
locket==0.2.1
logutils==0.3.5
lorem-text==2.1
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
more-itertools==8.10.0
mpmath==1.2.1
msgpack==1.0.2
multimethod==1.7
multipledispatch==0.6.0
murmurhash==1.0.6
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.8.0
netifaces==0.11.0
networkx==2.6.3
neural-structured-learning==1.3.1
nltk==3.6.3
notebook==6.4.4
numba==0.53.0
numexpr==2.7.3
numpy==1.22.4
oauthlib==3.1.1
object-detection==0.0.3
opencv-python==4.5.3.56
opencv-python-headless==4.1.2.30
openpyxl==3.0.9
opt-einsum==3.3.0
oslo.concurrency==5.0.1
oslo.config==9.0.0
oslo.context==5.0.0
oslo.i18n==5.1.0
oslo.log==5.0.1
oslo.messaging==14.0.0
oslo.metrics==0.5.0
oslo.middleware==5.0.0
oslo.serialization==5.0.0
oslo.service==3.0.0
oslo.utils==6.0.1
packaging==21.3
pamela==1.0.0
pandas==1.3.5
pandas-profiling==3.1.0
pandocfilters==1.5.0
panel==0.12.4
param==1.11.1
parso==0.8.2
partd==1.2.0
Paste==3.5.2
PasteDeploy==3.0.1
pathy==0.6.1
patsy==0.5.2
pbr==5.11.0
pdfminer==20191125
pecan==1.4.2
pexpect==4.8.0
phik==0.12.1
pickleshare==0.7.5
Pillow==8.3.2
pipreqs==0.4.11
platformdirs==2.4.0
plotly==5.3.1
pluggy==1.0.0
ply==3.11
pm4py==2.2.14
pm4pycvxopt==0.0.11
pooch==1.6.0
prefixspan==0.5.2
preshed==3.0.6
prettytable==3.5.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
proto-plus==1.20.1
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
PuLP==2.1
py==1.10.0
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.0
pycadf==3.1.1
pycairo==1.20.1
pycocotools==2.0.4
pycparser==2.20
pycryptodome==3.10.4
pyct==0.4.8
pycups==1.9.73
pycurl==7.44.1
pydantic==1.8.2
pydeck==0.7.1
pydot==1.4.2
pydotplus==2.0.2
PyGithub==1.55
Pygments==2.10.0
PyGObject==3.36.0
pygraphviz==1.7
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.1.0
pyLDAvis==3.3.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.4.0
pynndescent==0.5.4
pyod==0.9.4
pyOpenSSL==21.0.0
pyparsing==2.4.2
pyRFC3339==1.1
pyrsistent==0.18.0
PySimpleGUI==4.49.0
pysubgroup==0.7.2
pytest==6.2.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian==0.1.40
python-editor==1.0.4
python-json-logger==2.0.2
python-slugify==6.0.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyvis==0.1.9
pyviz-comms==2.1.0
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2021.9.30
repoze.lru==0.7
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
retrying==1.3.3
rfc3986==2.0.0
Routes==2.5.1
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.96
service-identity==21.1.0
shap==0.40.0
simplejson==3.17.5
six==1.14.0
sklearn==0.0
slicer==0.0.7
smart-open==5.2.1
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.3
SoundFile==0.10.3.post1
spacy==3.2.2
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SQLAlchemy==1.4.25
srsly==2.4.2
ssh-import-id==5.11
statsd==4.0.1
statsmodels==0.13.0
stevedore==4.1.1
streamlit==1.3.0
StringDist==1.0.9
suod==0.0.8
sympy==1.8
systemd-python==234
tangled-up-in-unicode==0.1.0
tb-nightly==2.9.0a20220218
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.9.1
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.6.0
tensorflow-model-optimization==0.7.1
tensorflowjs==3.13.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
testresources==2.0.1
text-unidecode==1.3
tf-models-official==2.3.0
tf-nightly-gpu==2.9.0.dev20220220
tf-slim==1.1.0
tflite-model-maker==0.3.4
tflite-support==0.3.1
thinc==8.0.13
threadpoolctl==3.0.0
tifffile==2021.8.30
toml==0.10.2
toolz==0.11.1
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
Twisted==21.7.0
typeguard==2.13.3
typer==0.4.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
umap-learn==0.5.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.11
validators==0.18.2
vine==5.0.0
virtualenv==20.8.1
visions==0.7.4
wadllib==1.3.6
wasabi==0.9.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.14.1
xarray==0.19.0
xes==1.3
xgboost==1.4.2
xkit==0.0.0
xlrd==2.0.1
yappi==1.4.0
yarg==0.1.9
zict==2.0.0
zipp==3.6.0
zope.interface==5.4.0
